Move Standalone related to shardingsphere-infra-standalone-mode module
