+++ shardingsphere-infra/shardingsphere-infra-mode/shardingsphere-infra-mode-type/shardingsphere-infra-standalone-mode/src/main/java/org/apache/shardingsphere/infra/mode/manager/standalone/repository/StandalonePersistRepositoryConfiguration.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.mode.standalone;
+package org.apache.shardingsphere.infra.mode.manager.standalone.repository;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
