+++ shardingsphere-infra/shardingsphere-infra-mode/shardingsphere-infra-mode-type/shardingsphere-infra-standalone-mode/src/main/resources/META-INF/services/org.apache.shardingsphere.infra.yaml.config.swapper.mode.PersistRepositoryConfigurationYamlSwapper
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.infra.yaml.config.swapper.mode.StandalonePersistRepositoryConfigurationYamlSwapper
+org.apache.shardingsphere.infra.mode.manager.standalone.yaml.swapper.StandalonePersistRepositoryConfigurationYamlSwapper
